Fix type conversion and casting errors across the codebase
- Fix decimal vs double comparisons (DepthBufferSelfTests.cs)
- Fix Interlocked.Read requiring long (ObjectPooling.cs, ConnectionHealthMonitor.cs)
- Fix double? to decimal? conversions (LockFreeOrderBook.cs)
- Add ForContext(Type) overload (LoggingSetup.cs)
- Add MarketEventType.Unknown enum value (MarketEventType.cs)
- Add MarketEventTier Tier property (MarketEvent.cs)
- Fix HistoricalBar timestamp access (TechnicalIndicatorService.cs)
- Change BackfillJobOptions to record, fix init-only properties (BackfillJob.cs)
- Add System.Reactive.Linq using directive (FallbackDataSourceOrchestrator.cs)
- Implement IDisposable for PluginLoadContext
- Fix nullable TResult return type (ConcurrentProviderExecutor.cs)
- Fix Symbol usage in Lean integration (SampleLeanAlgorithm.cs)
- Fix Histogram Buckets to use double[] (PrometheusMetrics.cs)

+++ MarketDataCollector/src/MarketDataCollector/Infrastructure/Performance/ObjectPooling.cs
@@ -16,7 +16,7 @@ public sealed class ObjectPool<T> where T : class
     private readonly Func<T> _factory;
     private readonly Action<T>? _reset;
     private readonly int _maxCapacity;
-    private int _currentCount;
+    private long _currentCount;
 
     public ObjectPool(Func<T> factory, Action<T>? reset = null, int maxCapacity = 256)
     {
@@ -62,7 +62,7 @@ public sealed class ObjectPool<T> where T : class
     /// <summary>
     /// Gets the current number of items in the pool.
     /// </summary>
-    public int Count => Interlocked.Read(ref _currentCount);
+    public int Count => (int)Interlocked.Read(ref _currentCount);
 }
 
 /// <summary>
